# packaging: use release builds and LTO by default