## Motivation

Let Meson handle optimization arguments for us.

Removed use of `debug` and `optimization` explicitly in project defaults -- these should be handled by Meson's `buildtype`, not us.

LTO is enabled when the build type is `release` or `minsize` and disabled otherwise.

Added links to relevant references describing Meson build types and the default build type as set by Nixpkgs' setup hook for Meson.

## Testing

I'm not sure what the state-of-the-art is for benchmarking Nix. Any pointers?

EDIT: I settled on gross rigging around `nix-functional-tests` to try to get a sense of the performance impact.

Step 1: Apply this patch so Meson doesn't randomly perturb malloc and runs tests sequentially: https://github.com/ConnorBaker/nix/commit/567a2b1699c5e5ecfd69fc75636dd4fedffcd123.

I've done that in the two branches I use for benchmarking: https://github.com/ConnorBaker/nix/tree/feat/det-bench for the baseline and https://github.com/ConnorBaker/nix/tree/feat/meson-O3-LTO-det-bench for this PR.

Of course, re-using the functional test suite as "micro-benchmarks" comes with its own problems...

Step 2: Decide on a tool for comparing benchmark numbers.

I decided on [`benchstat`](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) because it has the functionality I'm looking for, it is available in Nixpkgs, and the format for input data is relatively close to what we get as output from Meson (https://go.googlesource.com/proposal/+/master/design/14313-benchmark-format.md).

Step 3: Generate and transform the data.

Here's an AWK script which matches on the lines of the build output of `nix-functional-tests` which include the results of each script and transforms them into something which `benchstat` can work with. I named it `nix-functional-tests-to-gotest-format.awk`:

```awk
/.+ nix-functional-tests:.+ OK .+/ {
  suite=$3
  sub(/nix-functional-tests:/, "", suite)
  script=$5
  time=$7
  sub(/s/, "", time)
  print "BenchmarkTest" "/suite=" suite "/script=" script, "1", time, "s/op"
}
```

And here's a bash script which run the builds over and over, transforming the build logs and collecting them in output files. I named it `nix-functional-tests-gotest-bench.sh`:

```bash
#!/usr/bin/env bash
set -euo pipefail

# Old:
# https://github.com/ConnorBaker/nix/commits/feat/det-bench
# 567a2b1699c5e5ecfd69fc75636dd4fedffcd123

# New:
# https://github.com/ConnorBaker/nix/commits/feat/meson-O3-LTO-det-bench
# 10d5051bf93ea0cb732d8d41149b4f41fd049bdb

main() {
  local -A named_commits=(
    [old]="567a2b1699c5e5ecfd69fc75636dd4fedffcd123"
    [new]="10d5051bf93ea0cb732d8d41149b4f41fd049bdb"
  )
  local name=""
  local commit=""
  local flake_ref=""

  for name in "${!named_commits[@]}"; do
    commit="${named_commits[$name]}"
    flake_ref="github:ConnorBaker/nix/$commit"
    echo "Building the $name nix-functional-tests so dependencies are cached..."
    nix build -L --builders '' "$flake_ref#nix-functional-tests" --no-link
  done

  echo "Starting benchmarking..."

  for i in {1..20}; do
    echo "Running iteration $i/20..."
    for name in "${!named_commits[@]}"; do
      commit="${named_commits[$name]}"
      flake_ref="github:ConnorBaker/nix/$commit"
      echo "Running $name nix-functional-tests..."
      nix build -L --builders '' "$flake_ref#nix-functional-tests" --no-link --rebuild |& awk --file nix-functional-tests-to-gotest-format.awk >> "nix-functional-tests-$name.txt"
    done
  done
}

main
```

Step 4: Summarize the data.

I'm seeing (if I'm reading this right) roughly 5-10% improvements across most of the functional tests. This is on a NixOS 25.05 (`nixos-unstable`) system with an i9-13900k and `benchstat nix-functional-tests-old.txt nix-functional-tests-new.txt`:

<details>

```text
                                                               │ nix-functional-tests-old.txt │     nix-functional-tests-new.txt     │
                                                               │             s/op             │     s/op      vs base                │
Test/suite=main/script=test-infra                                                90.00m ± 11%   90.00m ±  0%    0.00% (p=0.008 n=20)
Test/suite=main/script=gc                                                        290.0m ±  3%   280.0m ±  4%   -3.45% (p=0.001 n=20)
Test/suite=main/script=nix-collect-garbage-d                                     340.0m ±  3%   325.0m ±  5%   -4.41% (p=0.000 n=20)
Test/suite=main/script=remote-store                                               1.785 ±  1%    1.630 ±  2%   -8.68% (p=0.000 n=20)
Test/suite=main/script=legacy-ssh-store                                          100.0m ±  0%   100.0m ± 10%    0.00% (p=0.003 n=20)
Test/suite=main/script=lang                                                       3.945 ±  1%    3.690 ±  1%   -6.46% (p=0.000 n=20)
Test/suite=main/script=characterisation-test-infra                               50.00m ±  0%   50.00m ±  0%        ~ (p=1.000 n=20)
Test/suite=main/script=experimental-features                                     345.0m ±  1%   330.0m ±  3%   -4.35% (p=0.000 n=20)
Test/suite=main/script=fetchMercurial                                             4.740 ±  0%    4.710 ±  0%   -0.63% (p=0.003 n=20)
Test/suite=main/script=gc-auto                                                    10.17 ±  0%    10.15 ±  0%   -0.20% (p=0.004 n=20)
Test/suite=main/script=user-envs                                                  1.360 ±  1%    1.250 ±  2%   -8.09% (p=0.000 n=20)
Test/suite=main/script=binary-cache                                               1.240 ±  2%    1.155 ±  2%   -6.85% (p=0.000 n=20)
Test/suite=main/script=multiple-outputs                                          495.0m ±  1%   460.0m ±  2%   -7.07% (p=0.000 n=20)
Test/suite=main/script=nix-build                                                 340.0m ±  3%   310.0m ±  3%   -8.82% (p=0.000 n=20)
Test/suite=main/script=gc-concurrent                                             240.0m ±  4%   240.0m ±  4%    0.00% (p=0.000 n=20)
Test/suite=main/script=repair                                                    505.0m ±  1%   470.0m ±  2%   -6.93% (p=0.000 n=20)
Test/suite=main/script=fixed                                                      7.410 ±  0%    7.380 ±  0%   -0.40% (p=0.000 n=20)
Test/suite=main/script=export-graph                                              500.0m ±  2%   450.0m ±  2%  -10.00% (p=0.000 n=20)
Test/suite=main/script=timeout                                                    6.200 ±  0%    6.190 ±  0%   -0.16% (p=0.000 n=20)
Test/suite=main/script=fetchGitRefs                                              935.0m ±  3%   865.0m ±  2%   -7.49% (p=0.000 n=20)
Test/suite=main/script=gc-runtime                                                 2.140 ±  0%    2.130 ±  0%   -0.47% (p=0.000 n=20)
Test/suite=main/script=tarball                                                  1020.0m ±  2%   955.0m ±  3%   -6.37% (p=0.000 n=20)
Test/suite=main/script=fetchGit                                                   1.275 ±  2%    1.210 ±  2%   -5.10% (p=0.000 n=20)
Test/suite=main/script=fetchurl                                                  340.0m ±  3%   320.0m ±  3%   -5.88% (p=0.000 n=20)
Test/suite=main/script=fetchPath                                                 60.00m ± 17%   60.00m ±  0%    0.00% (p=0.031 n=20)
Test/suite=main/script=fetchTree-file                                            150.0m ±  7%   140.0m ±  0%   -6.67% (p=0.000 n=20)
Test/suite=main/script=simple                                                    260.0m ±  4%   240.0m ±  4%   -7.69% (p=0.000 n=20)
Test/suite=main/script=referrers                                                 150.0m ±  7%   150.0m ±  7%        ~ (p=1.000 n=20)
Test/suite=main/script=optimise-store                                            170.0m ±  6%   170.0m ±  6%    0.00% (p=0.001 n=20)
Test/suite=main/script=substitute-with-invalid-ca                                150.0m ±  0%   140.0m ±  0%   -6.67% (p=0.000 n=20)
Test/suite=main/script=signing                                                   645.0m ±  4%   585.0m ±  1%   -9.30% (p=0.000 n=20)
Test/suite=main/script=hash-convert                                               1.340 ±  2%    1.165 ±  1%  -13.06% (p=0.000 n=20)
Test/suite=main/script=hash-path                                                1035.0m ±  2%   900.0m ±  2%  -13.04% (p=0.000 n=20)
Test/suite=main/script=gc-non-blocking                                            3.130 ±  2%    3.100 ±  3%   -0.96% (p=0.008 n=20)
Test/suite=main/script=check                                                     495.0m ±  3%   465.0m ±  1%   -6.06% (p=0.000 n=20)
Test/suite=main/script=nix-shell                                                 865.0m ±  3%   795.0m ±  2%   -8.09% (p=0.000 n=20)
Test/suite=main/script=check-refs                                                495.0m ±  1%   455.0m ±  1%   -8.08% (p=0.000 n=20)
Test/suite=main/script=build-remote-input-addressed                              645.0m ±  2%   590.0m ±  2%   -8.53% (p=0.000 n=20)
Test/suite=main/script=secure-drv-outputs                                        370.0m ±  3%   350.0m ±  3%   -5.41% (p=0.000 n=20)
Test/suite=main/script=restricted                                                485.0m ±  3%   450.0m ±  4%   -7.22% (p=0.000 n=20)
Test/suite=main/script=fetchGitSubmodules                                        700.0m ±  1%   670.0m ±  1%   -4.29% (p=0.000 n=20)
Test/suite=main/script=readfile-context                                          110.0m ±  0%   100.0m ±  0%   -9.09% (p=0.000 n=20)
Test/suite=main/script=nix-channel                                               470.0m ±  2%   430.0m ±  2%   -8.51% (p=0.000 n=20)
Test/suite=main/script=recursive                                                 310.0m ±  3%   280.0m ±  0%   -9.68% (p=0.000 n=20)
Test/suite=main/script=dependencies                                              330.0m ±  3%   310.0m ±  3%   -6.06% (p=0.000 n=20)
Test/suite=main/script=check-reqs                                                300.0m ±  3%   280.0m ±  4%   -6.67% (p=0.000 n=20)
Test/suite=main/script=build-remote-content-addressed-fixed                      340.0m ±  3%   310.0m ±  0%   -8.82% (p=0.000 n=20)
Test/suite=main/script=build-remote-content-addressed-floating                   340.0m ±  3%   320.0m ±  3%   -5.88% (p=0.000 n=20)
Test/suite=main/script=build-remote-trustless-should-pass-0                      180.0m ±  6%   170.0m ±  0%   -5.56% (p=0.000 n=20)
Test/suite=main/script=build-remote-trustless-should-pass-1                      180.0m ±  0%   170.0m ±  0%   -5.56% (p=0.000 n=20)
Test/suite=main/script=build-remote-trustless-should-pass-2                       5.250 ±  0%    5.240 ±  0%   -0.19% (p=0.000 n=20)
Test/suite=main/script=build-remote-trustless-should-pass-3                      190.0m ±  0%   175.0m ±  3%   -7.89% (p=0.000 n=20)
Test/suite=main/script=build-remote-trustless-should-fail-0                       5.160 ±  0%    5.150 ±  0%   -0.19% (p=0.000 n=20)
Test/suite=main/script=build-remote-with-mounted-ssh-ng                          120.0m ±  0%   120.0m ±  8%    0.00% (p=0.034 n=20)
Test/suite=main/script=nar-access                                                290.0m ±  3%   270.0m ±  4%   -6.90% (p=0.000 n=20)
Test/suite=main/script=impure-eval                                               180.0m ±  0%   160.0m ±  6%  -11.11% (p=0.000 n=20)
Test/suite=main/script=pure-eval                                                 250.0m ±  4%   230.0m ±  4%   -8.00% (p=0.000 n=20)
Test/suite=main/script=eval                                                      400.0m ±  3%   375.0m ±  4%   -6.25% (p=0.000 n=20)
Test/suite=main/script=repl                                                       2.000 ±  0%    1.950 ±  1%   -2.50% (p=0.000 n=20)
Test/suite=main/script=binary-cache-build-remote                                 230.0m ±  4%   220.0m ±  5%   -4.35% (p=0.000 n=20)
Test/suite=main/script=search                                                    295.0m ±  2%   280.0m ±  4%   -5.08% (p=0.000 n=20)
Test/suite=main/script=logging                                                   250.0m ±  4%   240.0m ±  4%   -4.00% (p=0.001 n=20)
Test/suite=main/script=export                                                    220.0m ±  0%   210.0m ±  5%   -4.55% (p=0.000 n=20)
Test/suite=main/script=config                                                    250.0m ±  4%   230.0m ±  4%   -8.00% (p=0.000 n=20)
Test/suite=main/script=add                                                       250.0m ±  4%   230.0m ±  4%   -8.00% (p=0.000 n=20)
Test/suite=main/script=chroot-store                                              150.0m ±  7%   140.0m ±  0%   -6.67% (p=0.000 n=20)
Test/suite=main/script=filter-source                                             90.00m ± 11%   80.00m ±  0%  -11.11% (p=0.025 n=20)
Test/suite=main/script=misc                                                      230.0m ±  4%   220.0m ±  5%   -4.35% (p=0.000 n=20)
Test/suite=main/script=dump-db                                                   160.0m ±  6%   145.0m ±  3%   -9.37% (p=0.000 n=20)
Test/suite=main/script=linux-sandbox                                             540.0m ±  2%   505.0m ±  1%   -6.48% (p=0.000 n=20)
Test/suite=main/script=supplementary-groups                                      180.0m ±  6%   180.0m ±  0%    0.00% (p=0.006 n=20)
Test/suite=main/script=build-dry                                                 260.0m ±  4%   240.0m ±  4%   -7.69% (p=0.000 n=20)
Test/suite=main/script=structured-attrs                                          270.0m ±  0%   250.0m ±  4%   -7.41% (p=0.000 n=20)
Test/suite=main/script=shell                                                     360.0m ±  0%   340.0m ±  3%   -5.56% (p=0.000 n=20)
Test/suite=main/script=brotli                                                    150.0m ±  0%   140.0m ±  0%   -6.67% (p=0.000 n=20)
Test/suite=main/script=zstd                                                      150.0m ±  0%   140.0m ±  0%   -6.67% (p=0.000 n=20)
Test/suite=main/script=compression-levels                                        125.0m ±  4%   120.0m ±  0%   -4.00% (p=0.002 n=20)
Test/suite=main/script=nix-copy-ssh                                              200.0m ±  5%   190.0m ±  5%   -5.00% (p=0.000 n=20)
Test/suite=main/script=nix-copy-ssh-ng                                           340.0m ±  3%   310.0m ±  3%   -8.82% (p=0.000 n=20)
Test/suite=main/script=post-hook                                                 280.0m ±  4%   260.0m ±  4%   -7.14% (p=0.000 n=20)
Test/suite=main/script=function-trace                                            130.0m ±  8%   120.0m ±  0%   -7.69% (p=0.000 n=20)
Test/suite=main/script=fmt                                                       150.0m ±  0%   140.0m ±  0%   -6.67% (p=0.000 n=20)
Test/suite=main/script=eval-store                                                230.0m ±  4%   210.0m ±  0%   -8.70% (p=0.000 n=20)
Test/suite=main/script=why-depends                                               180.0m ±  6%   170.0m ±  6%   -5.56% (p=0.000 n=20)
Test/suite=main/script=derivation-json                                           90.00m ±  0%   90.00m ±  0%    0.00% (p=0.023 n=20)
Test/suite=main/script=derivation-advanced-attributes                            70.00m ±  0%   70.00m ± 14%    0.00% (p=0.007 n=20)
Test/suite=main/script=import-from-derivation                                    290.0m ±  3%   260.0m ±  4%  -10.34% (p=0.000 n=20)
Test/suite=main/script=nix_path                                                  360.0m ±  3%   330.0m ±  3%   -8.33% (p=0.000 n=20)
Test/suite=main/script=nars                                                      470.0m ±  2%   430.0m ±  2%   -8.51% (p=0.000 n=20)
Test/suite=main/script=placeholders                                              70.00m ±  0%   70.00m ±  0%        ~ (p=1.000 n=20)
Test/suite=main/script=ssh-relay                                                100.00m ± 10%   90.00m ± 11%  -10.00% (p=0.000 n=20)
Test/suite=main/script=build                                                     610.0m ±  2%   570.0m ±  2%   -6.56% (p=0.000 n=20)
Test/suite=main/script=build-delete                                              380.0m ±  3%   360.0m ±  3%   -5.26% (p=0.000 n=20)
Test/suite=main/script=output-normalization                                      80.00m ± 12%   70.00m ±  0%  -12.50% (p=0.004 n=20)
Test/suite=main/script=selfref-gc                                                90.00m ±  0%   80.00m ±  0%  -11.11% (p=0.001 n=20)
Test/suite=main/script=bash-profile                                              40.00m ± 25%   40.00m ±  0%        ~ (p=0.176 n=20)
Test/suite=main/script=pass-as-file                                              70.00m ±  0%   70.00m ±  0%        ~ (p=0.184 n=20)
Test/suite=main/script=nix-profile                                                1.260 ±  2%    1.155 ±  3%   -8.33% (p=0.000 n=20)
Test/suite=main/script=suggestions                                               100.0m ±  0%   100.0m ±  0%        ~ (p=0.106 n=20)
Test/suite=main/script=store-info                                                70.00m ±  0%   65.00m ±  8%   -7.14% (p=0.002 n=20)
Test/suite=main/script=fetchClosure                                              310.0m ±  3%   290.0m ±  3%   -6.45% (p=0.000 n=20)
Test/suite=main/script=completions                                               310.0m ±  3%   290.0m ±  3%   -6.45% (p=0.000 n=20)
Test/suite=main/script=impure-derivations                                        430.0m ±  2%   400.0m ±  3%   -6.98% (p=0.000 n=20)
Test/suite=main/script=path-from-hash-part                                       80.00m ±  0%   80.00m ± 12%    0.00% (p=0.004 n=20)
Test/suite=main/script=path-info                                                 110.0m ±  0%   100.0m ± 10%   -9.09% (p=0.000 n=20)
Test/suite=main/script=toString-path                                             65.00m ±  8%   60.00m ±  0%   -7.69% (p=0.041 n=20)
Test/suite=main/script=read-only-store                                           200.0m ±  5%   180.0m ±  0%  -10.00% (p=0.000 n=20)
Test/suite=main/script=nested-sandboxing                                         755.0m ± 32%   880.0m ± 24%        ~ (p=0.784 n=20)
Test/suite=main/script=impure-env                                                820.0m ±  1%   810.0m ±  0%   -1.22% (p=0.030 n=20)
Test/suite=main/script=debugger                                                  80.00m ±  0%   80.00m ± 12%        ~ (p=0.052 n=20)
Test/suite=libstoreconsumer/script=test-libstoreconsumer                         80.00m ± 12%   80.00m ±  0%    0.00% (p=0.001 n=20)
Test/suite=plugins/script=plugins                                                50.00m ± 20%   50.00m ±  0%    0.00% (p=0.020 n=20)
Test/suite=ca/script=build-with-garbage-path                                     180.0m ±  6%   170.0m ±  6%   -5.56% (p=0.000 n=20)
Test/suite=ca/script=build                                                       565.0m ±  1%   520.0m ±  2%   -7.96% (p=0.000 n=20)
Test/suite=ca/script=build-cache                                                 460.0m ±  2%   430.0m ±  2%   -6.52% (p=0.000 n=20)
Test/suite=ca/script=concurrent-builds                                            10.07 ± 50%    10.07 ± 50%        ~ (p=0.413 n=20)
Test/suite=ca/script=derivation-json                                             150.0m ±  0%   140.0m ±  7%   -6.67% (p=0.001 n=20)
Test/suite=ca/script=duplicate-realisation-in-closure                             2.165 ±  0%    2.155 ±  0%   -0.46% (p=0.000 n=20)
…