@Ericson2314 2024-12-21
Won't this slow down debug builds?