@MagicRB 2024-12-21
This should be 100% disabled for debug builds, optimization can make chabges to the code which make debugging really hard, inlining, the infamoud "where did my code go? right it was optimized away/unrolled"