@ConnorBaker 2024-12-21
Should debug builds be built without any optimization? Because that’s not what happens currently :/