@MagicRB 2024-12-21
I'm not part of the Nix dev team, but in my experience having all optimizations off greatly helps starring at code in gdb/lldb. Since the machine code is closer to what's actually written the source code