@Ericson2314 2024-12-24
I think without any optimization, right @edolstra?

We could have the full nix builds instead specify more optimization with `mesonFlags` I think?