@Mic92 2024-12-25
> @Ericson2314 so should debug builds happen without any optimization, or with optimization level two (which is what happens prior to this PR)?

We should use `-Og` rather than `-O0`