@ConnorBaker 2024-12-29
Updated and force-pushed.

Removed use of `debug` and `optimization` explicitly in project defaults -- these should be handled by Meson's `buildtype`, not us.

LTO is enabled when the build type is `release` or `minsize` and disabled otherwise.

Added links to relevant references describing Meson build types and the default build type as set by Nixpkgs' setup hook for Meson.