@ConnorBaker 2025-01-01
LTO on Darwin was broken until fairly recently, so that explains that.
I guess I conditionally disable it and make a note.