@ConnorBaker 2025-01-01
> @ConnorBaker thanks for your work, I like the looks of this a lot now. Yeah better do that macOS disable for now (unless you know how to fix) but then let's merge it!

Fixing it would require a bump to our pin of Nixpkgs — is it safe to say that would need to be a separate PR with more testing?