@ConnorBaker 2025-01-02
I rebased and force-pushed; I think the locked version master is using for `release-24.11` is new enough.
I’m re-running the benchmarks (including on my Mac, we’ll see how that goes) right now!