@ConnorBaker 2025-01-02
Updated the benchmarks; performance gain on my i9-13900k is basically unchanged (still 5-10%) and on macOS I'm seeing a 1-2% gain.

@Ericson2314 I'm happy with this now and CI is green; anything else we need to do?