@ConnorBaker 2025-01-02
@Radvendii I remember we had discussed a benchmark suite for Nix at some point as a way to establish a baseline and collect metrics; you may be interested in what I in the PR description. It's super gross, but that sort of functionality would be really, really helpful (and avoid people trying to roll their own suites for performance improving PRs)!